@@ -7,31 +7,33 @@
     "commit-release": "dist/bin.js"
   },
   "bugs": "https://github.com/JamieMason/commit-release/issues",
-  "config": {
-    "commitizen": {
-      "path": "./node_modules/cz-conventional-changelog"
-    }
-  },
   "dependencies": {
-    "chalk": "2.4.2",
-    "commander": "2.20.0",
-    "conventional-changelog-cli": "2.0.21",
-    "conventional-recommended-version": "0.5.2",
-    "execa": "2.0.3",
-    "fs-extra": "8.1.0",
-    "package-json-to-readme": "2.1.1",
-    "prettier": "1.18.2"
+    "chalk": "4.1.2",
+    "commander": "10.0.0",
+    "conventional-recommended-version": "1.1.0",
+    "fs-extra": "11.1.0",
+    "giterator": "0.6.0",
+    "prettier": "2.8.4"
   },
   "devDependencies": {
+    "@tsconfig/node14": "1.0.3",
     "@types/chalk": "2.2.0",
     "@types/commander": "2.12.2",
-    "@types/execa": "2.0.0",
-    "@types/fs-extra": "8.0.0",
-    "@types/node": "12.6.8",
-    "@types/prettier": "1.18.0",
-    "rimraf": "2.6.3",
-    "tslint": "5.18.0",
-    "typescript": "3.5.3"
+    "@types/fs-extra": "11.0.1",
+    "@types/jest": "29.4.0",
+    "@types/node": "14.18.36",
+    "@types/prettier": "2.7.2",
+    "@typescript-eslint/eslint-plugin": "5.53.0",
+    "@typescript-eslint/parser": "5.53.0",
+    "eslint": "8.34.0",
+    "eslint-plugin-import": "2.27.5",
+    "eslint-plugin-jest": "27.2.1",
+    "expect-more-jest": "5.5.0",
+    "jest": "29.4.3",
+    "rimraf": "4.1.2",
+    "ts-jest": "29.0.5",
+    "ts-node": "10.9.1",
+    "typescript": "4.9.5"
   },
   "files": [
     "dist"
@@ -53,15 +55,15 @@
   "main": "dist/index.js",
   "repository": "JamieMason/commit-release",
   "resolutions": {
-    "lodash": ">=4.17.13"
+    "chalk": "4.1.2"
   },
   "scripts": {
-    "build": "tsc --project .",
-    "format": "prettier --write '**/*.md' './src/**/*.ts'",
-    "lint": "tslint --fix --project ./tsconfig.json",
-    "prebuild": "rimraf ./dist",
-    "prelint": "npm run format",
-    "prepack": "npm run build"
-  },
-  "typings": "dist/index.d.ts"
+    "build": "rm -rf ./dist && tsc --project .",
+    "format": "yarn format:lint && yarn format:source",
+    "format:lint": "yarn lint --fix",
+    "format:source": "prettier --write .",
+    "lint": "eslint --ext .ts .",
+    "prepack": "yarn build",
+    "test": "jest src test"
+  }
 }
